+++ templates/oop_rails_server_base/app/controllers/working_controller.rb
@@ -1,5 +1,5 @@
 class WorkingController < ApplicationController
   def rails_is_working
-    render :text => "Rails version: #{Rails.version}\nRuby version: #{RUBY_VERSION}"
+    render :text => "Rails version: #{Rails.version}\nRuby version: #{RUBY_VERSION}\nRuby engine: #{RUBY_ENGINE}\n"
   end
 end